Memoizing typecasts cuts parsing time in half
When parsing a file with one root element and ~1000 child elements with each a dozen attributes, this patch cuts processing time on my machine in half. It's a traditional space-time tradeoff. The time cost is reduced because the garbage collector has much less to do.

Note that this is not optimal for inputs with little duplication in their primitive values and could be problematic for large inputs with little duplication in their primitive values. However, the latter is quite uncommon.