Emit attributes even when a default value is specified
HappyMapper's current behavior does not allow support for XSDs that specify default values for attributes while also specifying  `use="required"`, meaning they must always be present in XML messages conforming to the schema. Some w3c XSD's use this combination. Since HappyMapper does not support specifying whether an attribute is required, you currently need a workaround to ensure such an attribute is always emitted.

To change this, we cannot support the optimization of not emitting attributes that have a default value. To support that again, Attribute would need a `use`, `optional` or `required` option to specify the desired behavior.