Emit attributes when state_when_nil is true, independent of value
Attributes with option `:state_when_nil` set are intended to have special behavior when their value is nil: they should be emitted anyway. However, the previous code never did so: the first term of the if clause would evaluate to true and the second term would never be evaluated.

This commit changes it so the attribute is actually emitted when the value is nil and `:state_when_nil` is set.