Allow explicitly blocking namespace prefixes on model child elements
Setting namespace to nil on a model child element was equivalent to not
setting it, both for reading and writing. This commit changes that
by using an empty namespace when searching for nodes to parse, and by
the namespace on the generated element.

Part of https://github.com/mvz/happymapper/issues/43.